requirements.txt
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0

opencv-python>=4.10.0
numpy>=1.24.0
scipy>=1.10.0

pytesseract>=0.3.10
Pillow>=10.0.0

matplotlib>=3.7.0
pandas>=2.0.0

streamlit>=1.35.0

tqdm>=4.66.0

